+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Reject HTTP signature if key ID doesn't match one indicated by `publicKey` property.
 - Add `emoji` and `emoji_url` fields to `Notification` entity for compatibility with Pleroma.
 - Address `Move` activity to `Public` and followers collection instead of individual actors.
+- Do not allow more than 10 identity proofs.
+- Do not allow more than 10 aliases.
 
 ### Removed